fix: resolve build errors with KeyBinding.boundKey and type mismatch
- Replace direct access to private KeyBinding.boundKey field with
  KeyBindingHelper.getBoundKeyOf() from Fabric API (10 occurrences in
  ActionManager.kt)
- Fix Long to Int type mismatch in BotCore.kt by adding .toInt()
  conversion for getAdaptiveDelay() return value

+++ src/client/kotlin/fr/nix/agribot/bot/BotCore.kt
@@ -523,7 +523,7 @@ object BotCore {
             waitMs(200)
             BucketManager.pourWaterBucket()
             // Utiliser le delai adaptatif pour le prochain seau
-            waitMs(BucketManager.getAdaptiveDelay())
+            waitMs(BucketManager.getAdaptiveDelay().toInt())
         }
 
         // Limite de securite
